Bump to newer vert.x versions.
This only takes us up to CR3 for vertx-core & vertx-platform. The
integration tests fail if we build with 2.0.0.final due to classloader
issues, but a module built against CR3 will work fine with final.

I need to chat with Tim about possibly changing testtools to work
around the CL issues.