merge tobias
This documents most things, and is a port of the ruby lang module
manual. The sockjs section is all that remains to be ported.

Remove unused import.

Allow passing a filtering function for the event.

Whitespace.

Move write to stream ns, since that's really what you're operating on.

Add method functions.

merge some change with tobias master